+++ tests/spelling.R
@@ -1,3 +1,1 @@
-if (identical(Sys.getenv("NOT_CRAN"), "true")) {
-  spelling::spell_check_test(vignettes = TRUE, error = FALSE)
-}
+spelling::spell_check_test(vignettes = TRUE, error = FALSE)